requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
black==23.3.0
certifi==2023.5.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.2.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-spectacular==0.26.2
et-xmlfile==1.1.0
exceptiongroup==1.1.1
filelock==3.12.0
frozenlist==1.4.0
identify==2.5.24
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.8.0
numpy==1.24.3
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
platformdirs==3.5.1
pluggy==1.0.0
pre-commit==3.3.2
psycopg2-binary==2.9.6
pyrsistent==0.19.3
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
virtualenv==20.23.0
yarl==1.9.2
